src/functions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure.functions
azure-functions
flask==2.2.5
flask-cors==3.0.10
azure-common==1.1.3
azure-nspkg==1.0.0
azure-storage==0.31.0
azure-storage-blob==12.15.0
azure-mgmt-storage==21.0.0
azure-cognitiveservices-speech==1.26.0
openai
tiktoken==0.3.0
